requirements.txt
fastapi==0.115.8
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.6.0
uvicorn==0.34.0
pytest
httpx
